Created ITransaction.ts, configured useAxios for TransactionsContainer
