+++ ui/src/pages/plan/transactions/ITransaction.ts
@@ -1,0 +1,11 @@
+
+export interface IApiTransaction {
+    rule_id: string;
+    id: string;
+    value: number;
+    day: string;
+    calculations: {
+        balance: number;
+        working_capital: number;
+    }
+}
